fix: move create_indexing_tasks to indexing_task.py
